// chpqa-app/src/app/tasks/form-submission/model/form-submission-path.model.ts
export enum FormSubmissionPath {
  BASE_PATH = '/form-submission',
  TASK_LIST = 'task-list',
  REVIEW_ADDRESS_AND_SITE_CONTACT = 'review-address-and-site-contact',
  UPLOAD_SCHEME_LINE_DIAGRAM = 'upload-scheme-line-diagram',
  SCHEME_LINE_DIAGRAM_SUMMARY = 'scheme-line-diagram-summary',
  UPLOAD_ENERGY_FLOW_DIAGRAM = 'upload-energy-flow-diagram',
  ENERGY_FLOW_DIAGRAM_SUMMARY = 'energy-flow-diagram-summary',
  UPLOAD_DAILY_HEAT_PROFILE = 'upload-daily-heat-profile',
  DAILY_HEAT_PROFILE_SUMMARY = 'daily-heat-profile-summary',
  UPLOAD_ANNUAL_HEAT_PROFILE = 'upload-annual-heat-profile',
  ANNUAL_HEAT_PROFILE_SUMMARY = 'annual-heat-profile-summary',
  UPLOAD_HEAT_LOAD_DURATION_CURVE = 'upload-heat-load-duration-curve',
  HEAT_LOAD_DURATION_CURVE_SUMMARY = 'heat-load-duration-curve-summary',
  SET_TOTAL_CAPACITY_UNDER_MAXHEAT_CONDITIONS = 'set-total-power-capacity-under-maxheat',
  TOTAL_CAPACITY_UNDER_MAXHEAT_CONDITIONS_SUMMARY = 'total-power-capacity-under-maxheat-summary',
  ADD_PRIME_MOVER = 'add-prime-mover',
  PRIME_MOVER_MECHANICAL_LOAD = 'prime-mover-mechanical-load',
  PRIME_MOVER_ENGINE_TYPE = 'prime-mover-engine-type',
  PRIME_MOVER_ENGINE_SUBTYPE = 'prime-mover-engine-subtype',
  PRIME_MOVER_ENGINE_MANUFACTURER = 'prime-mover-engine-manufacturer',
  PRIME_MOVER_ENGINE_MODEL = 'prime-mover-engine-model',
  PRIME_MOVER_ENGINE = 'prime-mover-engine',
  PRIME_MOVER_FIGURES = 'prime-mover-figures',
  PRIME_MOVER_DOCUMENTATION = 'prime-mover-documentation',
  PRIME_MOVER_REVIEW_ANSWERS = 'prime-mover-review-answers',
  PRIME_MOVER_CONFIRM_DELETION = 'prime-mover-confirm-deletion',
  ADD_METER = 'add-meter',
  METER_EXISTENCE = 'meter-existence',
  METER_OUTPUT_RANGE = 'meter-output-range',
  METER_TYPE = 'meter-type',
  METER_FISCAL_CHECK = 'meter-fiscal-check',
  METER_UNCERTAINTY = 'meter-uncertainty',
  METER_MEASUREMENT = 'meter-measurement',
  METER_DOCUMENTATION = 'meter-documentation',
  METER_REVIEW_ANSWERS = 'meter-review-answers',
  METER_CONRFIRM_DELETION = 'meter-confirm-deletion',
  PROVIDE_HOURS_OF_OPERATION = 'provide-hours-of-operation',
  PROVIDE_HOURS_OF_OPERATION_SUMMARY = 'provide-hours-of-operation-summary',
  PROVIDE_ENERGY_INPUTS = 'provide-energy-inputs',
  ENTER_ENERGY_INPUTS = 'enter-energy-inputs',
  PROVIDE_POWER_OUTPUTS = 'provide-power-outputs',
  ENTER_POWER_OUTPUT = 'enter-power-output',
  POWER_EFFICIENCY_THRESHOLD = 'power-efficiency-threshold',
  PROVIDE_HEAT_OUTPUTS = 'provide-heat-outputs',
  ENTER_HEAT_OUTPUT = 'enter-heat-output',
  CONDESING_STEAM_TURBINE = 'provide-condensing-steam-turbine-details',
  CONDESING_STEAM_TURBINE_SUMMARY = 'condensing-steam-turbine-summary',
  QUALITY_INDEX_STATUS = 'quality-index-status',
  ROC_QUALITY_INDEX_STATUS = 'renewables-obligation-certificate-quality-index-status',
  CFD_QUALITY_INDEX_STATUS = 'contracts-for-difference-certificate-quality-index-status',
  REQUEST_SECRETARY_OF_STATE_EXEMPTION_CERTIFICATE = 'secretary-of-state-exemption-certificate',
  SECRETARY_OF_STATE_EXEMPTION_CERTIFICATE_SUMMARY = 'secretary-of-state-exemption-certificate-summary',
  REQUEST_ROC_CERTIFICATE = 'renewables-obligation-certificate',
  ROC_CERTIFICATE_SUMMARY = 'renewables-obligation-certificate-summary',
  REQUEST_CFD_CERTIFICATE = 'request-contracts-for-difference-certificate',
  CFD_CERTIFICATE_SUMMARY = 'contracts-for-difference-certificate-summary',
  FINANCIAL_BENEFITS = 'provide-financial-benefits-information',
  FINANCIAL_BENEFITS_SUMMARY = 'financial-benefits-summary',
  QUALITY_INDEX_THRESHOLD = 'quality-index-threshold',
  QUALITY_INDEX_THRESHOLD_SUMMARY = 'quality-index-threshold-summary',
  SUBMIT_FORM = 'submit-form',
  SUBMIT_SUCCESS = 'submit-success',

  PROVIDE_UNCERTAINTY_FACTORS = 'provide-uncertainty-factors',
  PROVIDE_UNCERTAINTY_FACTORS_UPLOAD = 'upload-uncertainty-factors',
  PROVIDE_UNCERTAINTY_FACTORS_SUMMARY = 'provide-uncertainty-factors-summary',

  UPLOAD_PERFORMANCE_DATA = 'upload-performance-data',
  UPLOAD_PERFORMANCE_DATA_SUMMARY = 'upload-performance-data-summary',
  UPLOAD_PERFORMANCE_DATA_SUBMIT_PROGRESS = 'upload-performance-data-submit-progress',

  HEAT_REJECTION_FACILITY = 'heat-rejection-facility',
  HEAT_REJECTION_FACILITY_SUMMARY = 'heat-rejection-facility-summary',

  SCHEME_HISTORY = 'scheme-history'
}